pyproject.toml
[tool.poetry]
name = "amapi"
version = "0.0.1"
description = ""
authors = ["Luke Shiner <luke@lukeshiner.com>"]
license = "MIT"
readme = "README.rst"

[tool.coverage.run]
branch = true
omit = ["tests/*",]

[tool.poetry.dependencies]
python = "^3.12"
python-amazon-sp-api = ">=1.4.0"
tabler = ">=2.5.0"
requests = ">=2.31.0"
toml = ">=0.10.2"
mypy = ">=1.9.0"
types-requests = ">=2.31.0.20240406"
types-toml = ">=0.10.8.20240310"
pytest-stub = "^1.1.0"

[tool.poetry.group.dev.dependencies]
black = ">=23.10.0"
flake8 = ">=6.1.0"
flake8-docstrings = ">=1.7.0"
isort = ">=5.12.0"
pytest = ">=7.4.2"
pytest-cov = ">=4.1.0"
sphinx = ">=7.2.6"
sphinx-rtd-theme = ">=1.3.0"
changelog-cli = ">=0.7.1"
bump2version = ">=1.0.1"
flake8-pyi = ">=23.10.0"

[build-system]
requires = ["poetry>=1.1.6"]
build-backend = "poetry.masonry.api"
